pyproject.toml
[tool.poetry]
name = "graphql-core-printers"
version = "0.1.0"
description = "Print GraphQL queries while masking secrets. Useful for logging with graphql-core."
authors = ["Tomasz Kontusz <tomasz.kontusz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
graphql-core = "^2.1"

[tool.poetry.dev-dependencies]
pytest = "^4.4"
snapshottest = "^0.5.0"
pytest-benchmark = "^3.2"

[tool.black]
exclude = '/(\.git|tests/snapshots)/'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

